# fixes for PR #280, refactoring, textlogger, unit test