**What this PR does / why we need it**:

https://github.com/kubernetes/klog/pull/280/ turned println and printf into
wrapper functions, but the underlying *Depth functions kept passing the same 0
depth to l.output and thus any logr.Logger. That off-by-one error broke JSON tests in Kubernetes, luckily well before any of the PR went into a release. 

There was also some confusion about println vs print.

This shows that we need better unit test coverage, in particular for output in combination with SetLogger. We cannot depend on some external logger for that. This PR solves that by splitting out all the relevant code in klog into internal helper packages and then implementing a Logger that produces output the same way as klog.

That makes it possible to write a test that can be applied to both klog, klog+logger, and logger without klog. The latter is something that may become relevant for contextual logging in Kubernetes: if we ever decided to stop logging through klog, then such a stand-alone logger can provide the implementation for logging in text format.

The unit test is exposed with an API that allows out-of-tree Loggers to reuse the test cases. This has been tested with the JSON logger in Kubernetes already and will be used there once this PR is in a release.

**Which issue(s) this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*:

Fixes https://github.com/kubernetes/klog/pull/280/files#r800965182
Fixes https://github.com/kubernetes/klog/pull/280/files#r803909527
Fixes: https://github.com/kubernetes/klog/issues/288

**Release note**:
```release-note
NONE
```