@pohly 2022-02-10
Additional commits for contextual logging are based on this PR and currently can be found in https://github.com/pohly/klog/commits/contextual-logging