@pohly 2022-02-11
This refactoring has no performance impact:
```
name                   old time/op  new time/op  delta
Header-36              1.63µs ± 8%  1.64µs ± 5%   ~     (p=1.000 n=5+5)
HeaderWithDir-36       1.65µs ± 7%  1.67µs ± 3%   ~     (p=0.548 n=5+5)
V-36                   2.44ns ± 3%  2.42ns ± 2%   ~     (p=0.310 n=5+5)
KRef-36                0.50ns ± 0%  0.50ns ± 0%   ~     (p=0.413 n=5+4)
KObj-36                6.59ns ± 0%  6.59ns ± 0%   ~     (p=0.556 n=5+4)
Logs-36                6.73µs ± 4%  6.87µs ± 9%   ~     (p=0.690 n=5+5)
WithoutDeferUnLock-36  12.6ns ± 3%  12.6ns ± 1%   ~     (p=0.444 n=5+5)
WithDeferUnLock-36     13.6ns ± 1%  13.6ns ± 1%   ~     (p=0.722 n=5+5)
```

I had tried to use the `internal/verbosity` package also in klog.go, but that had made `V()` slower, so I left the original code in place unmodified.