@pohly 2022-02-11
> Maybe one note, it's cleaner to put the commit that adds tests at the first one instead ot the last.

I see what you mean. I've rearranged the commits as you suggested, with modified tests to make "go test" pass for each commit.

One exception is testing `SetLogger` with the text logger: that has to be added after fixing the stack unwinding bug because the tests can't be modified to pass while that bug is present.

> It's puts some additional work to update them,

Was too bad.

>  however it's worth to give reviewers a trace of what is changing in each commit.

Agreed. Perhaps take another look?