@pohly 2022-02-11
> One thing that we need to clarify: some of the new API should be considered subject to change ((like the entire "test" package) or alpha (text logger). How do we communicate that so that if we need to break their API, we are not bound by the v2 API stability guarantee?

@dims: any thoughts on that?