@dims 2022-02-11
@pohly the only thing i can think of is to add a README stating things explicitly. let's poke @liggitt @thockin on this to see if they have any other ideas.