@pohly 2022-02-11
A section in the README explaining stability guarantees seems reasonable. I would also call it out explicitly in the doc comment, to make sure people see it also in pkg.go.dev and `go doc`. I'm just not sure whether there are already any conventions for this... I found this:

https://github.com/grpc/grpc-go/blob/b4c5e24c7373b22c6c38d68379c39ecdc616a968/server.go#L1665-L1681
https://pkg.go.dev/google.golang.org/grpc#NewContextWithServerTransportStream

This looks suitable.