@pohly 2022-02-14
I've added "EXPERIMENTAL" docs comments and a section to the README explaining this.

@dims: I think this PR is ready now. Also, can you perhaps set a `tools/v0.1.0` tag? I want to check whether that really allows importing the logcheck tool into Kubernetes with a version instead of a commit hash.